requirements.txt
alembic==1.14.0
appdirs==1.4.3
apturl==0.5.2
bcrypt==3.1.7
blinker==1.8.2
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.1.7
colorama==0.4.3
command-not-found==0.3
cryptography==35.0.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
ecdsa==0.19.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flask==3.0.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
flask-talisman==0.8.1
future==0.18.2
greenlet==3.1.1
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pbr==6.1.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.3.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.19.1
python-jose-cryptodome==1.3.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
SQLAlchemy==2.0.36
systemd-python==234
testresources==2.0.1
typing-extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
werkzeug==3.0.6
xkit==0.0.0
zipp==3.20.2
